[value-path-wiring] add value-path correctness tests + C-1 idempotency-key regression guard
Per-user custody signing (user key, IDOR reject), crash-replay exactly-once
(broadcast+complete crash → no re-mint, reconciliation settles from the
persisted key), KYC at the raw mint door (no upsert/op for unverified avatar),
no-truncation on >int.MaxValue amounts, M1 timeout preserves txId. Plus the
red-green AmountIdempotencyKeyTests pinning C-1: the historic int clamp collapsed
two distinct >int.MaxValue amounts to one idempotency key (silent
under-allocation); the ulong projection keeps them distinct, and the real
DeriveIdempotencyKey path now yields 2 distinct keys + 2 broadcasts.

+++ tests/OASIS.WebAPI.Tests/Signing/AlgorandProviderTransactTests.cs
@@ -129,7 +129,7 @@ public class AlgorandProviderTransactTests : IDisposable
     public async Task Transfer_builds_signs_and_submits()
     {
         const ulong tokenId = 12345UL;
-        const int transferAmount = 1;
+        const ulong transferAmount = 1UL;
         using var _ = RunStub(confirmedRound: 3, assetIndex: null, poolError: null);
 
         var provider = NewProvider();
